llm/gemma-embd/pyproject.toml
[project]
name = "gemma-embd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "protobuf>=7.34.0",
    "requests>=2.32.5",
    "sentencepiece>=0.2.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
]
